Move components to proper directory
