+++ src/services/ui/src/components/PackageDetails/DetailItemsGrid.tsx
